requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.13.0
matplotlib>=3.5.0
seaborn>=0.12.0
wordcloud>=1.8.0
nltk>=3.8.0
scikit-learn>=1.1.0
# New dependencies for advanced features
bertopic>=0.14.0
networkx>=2.8.0
python-dotenv>=0.21.0
google-generativeai>=0.3.0
hdbscan>=0.8.29
umap-learn>=0.5.3
sentence-transformers>=2.2.0
requests>=2.28.0
